payment/requirements.txt
Flask==3.0.2
redis==5.0.3
gunicorn==21.2.0
msgspec==0.18.6
requests==2.32.0
grpcio~=1.71.0
fastapi
uvicorn
dotenv